Fix: divide webhook tests into separate files
